telegram-tracker/requirements.txt
python-telegram-bot[job-queue]==21.6
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
openpyxl==3.1.5
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
